コメントの修正　HealthMetricCardUiをstateファイルに移動　typeをRecordTypeから取得
