+++ app/src/main/java/com/example/myaku_rismu/feature/home/HomeScreen.kt
@@ -38,9 +38,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.myaku_rismu.R
 import com.example.myaku_rismu.core.AppState
+import com.example.myaku_rismu.data.model.RecordType
 import com.example.myaku_rismu.feature.home.components.BarChart
 import com.example.myaku_rismu.feature.home.components.DonutChart
 import com.example.myaku_rismu.feature.home.components.GifImageLoader
@@ -50,14 +50,6 @@ import com.example.myaku_rismu.ui.theme.Myaku_rismuTheme
 import com.example.myaku_rismu.ui.theme.customTheme
 import kotlinx.coroutines.launch
 
-data class HealthMetricCardUi(
-    val title: Int,
-    val genre: Int,
-    val unit: Int,
-    val icon: Int,
-    val color: Color,
-    val barColorFaded: Color,
-)
 
 @Composable
 fun HomeScreen(
@@ -88,10 +80,6 @@ fun HomeScreen(
 
             is HomeUiEvent.ShowBottomSheet -> {
                 viewModel.showBottomSheet()
-            }
-
-            is HomeUiEvent.SelectHealthMetric -> {
-                appState.navigateToHealthDetail(metricType = event.type)
             }
 
             is HomeUiEvent.CreateNewMusic -> {
@@ -167,6 +155,7 @@ fun HomeContent(
 ) {
     val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
     val scope = rememberCoroutineScope()
+
     HomeBottomSheet(
         show = uiState.showBottomSheet,
         sheetState = sheetState,
@@ -200,9 +189,7 @@ fun HomeContent(
                 cardList = cardList,
                 modifier = Modifier
                     .padding(start = 16.dp, top = 16.dp, end = 16.dp, bottom = 4.dp),
-                onClick = { metric ->
-                    eventHandler(HomeUiEvent.SelectHealthMetric(metric.type.toString()))
-                }
+                onClick = {}
             )
         }
     }
@@ -276,7 +263,7 @@ fun HealthMetricCard(
     modifier: Modifier = Modifier,
     metric: HealthMetric,
     cardUi: HealthMetricCardUi,
-    onClick: () -> Unit = { /* No-op by default */ }
+    onClick: () -> Unit = {}
 ) {
     Card(
         modifier = modifier,
@@ -385,15 +372,15 @@ fun HealthMetricsSection(
     uiState: HomeState,
     cardList: List<HealthMetricCardUi>,
     modifier: Modifier = Modifier,
-    onClick: (HealthMetric) -> Unit = { /* No-op by default */ }
+    onClick: (HealthMetric) -> Unit = {}
 ) {
     Column(
         modifier = modifier,
         verticalArrangement = Arrangement.spacedBy(9.dp)
     ) {
-        // 1つ目
         val firstMetric = uiState.metrics[0]
         val firstCardUi = cardList.getOrNull(0)
+
         if (firstCardUi != null) {
             HealthMetricCard(
                 modifier = Modifier
@@ -411,6 +398,7 @@ fun HealthMetricsSection(
             ) {
                 rowItems.forEachIndexed { colIndex, metric ->
                     val cardUi = cardList.getOrNull(rowIndex * 2 + colIndex + 1)
+
                     if (cardUi != null) {
                         HealthMetricCard(
                             modifier = Modifier
@@ -433,27 +421,27 @@ fun HomeScreenPreview() {
     val dummyUiState = HomeState(
         metrics = listOf(
             HealthMetric(
-                type = HomeHealthType.HEART_RATE,
+                type = RecordType.HEART_RATE,
                 currentValue = 200,
                 targetValue = 180
             ),
             HealthMetric(
-                type = HomeHealthType.STEPS,
+                type = RecordType.STEPS,
                 currentValue = 5000,
                 targetValue = 10000
             ),
             HealthMetric(
-                type = HomeHealthType.CALORIES,
+                type = RecordType.CALORIES,
                 currentValue = 1200,
                 targetValue = 2000
             ),
             HealthMetric(
-                type = HomeHealthType.SLEEP_TIME,
+                type = RecordType.SLEEP_TIME,
                 currentValue = 9,
                 targetValue = 8
             ),
             HealthMetric(
-                type = HomeHealthType.DISTANCE,
+                type = RecordType.DISTANCE,
                 currentValue = 2,
                 targetValue = 5
             )
@@ -501,6 +489,7 @@ fun HomeScreenPreview() {
             barColorFaded = MaterialTheme.customTheme.homeMoveDistanceBarColorFaded,
         )
     )
+
     Myaku_rismuTheme {
         HomeContent(
             uiState = dummyUiState,
