+++ app/src/main/java/com/example/myaku_rismu/feature/home/HomeViewModel.kt
@@ -8,6 +8,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 import com.example.myaku_rismu.core.ScreenState
+import com.example.myaku_rismu.data.model.RecordType
 import kotlinx.coroutines.flow.update
 
 @HiltViewModel
@@ -65,27 +66,27 @@ class HomeViewModel @Inject constructor(
         // 仮データ
         val metrics = listOf(
             HealthMetric(
-                type = HomeHealthType.HEART_RATE,
+                type = RecordType.HEART_RATE,
                 currentValue = 200,
                 targetValue = 180
             ),
             HealthMetric(
-                type = HomeHealthType.STEPS,
+                type = RecordType.STEPS,
                 currentValue = 5000,
                 targetValue = 10000
             ),
             HealthMetric(
-                type = HomeHealthType.CALORIES,
+                type = RecordType.CALORIES,
                 currentValue = 1200,
                 targetValue = 2000
             ),
             HealthMetric(
-                type = HomeHealthType.SLEEP_TIME,
+                type = RecordType.SLEEP_TIME,
                 currentValue = 9,
                 targetValue = 8
             ),
             HealthMetric(
-                type = HomeHealthType.DISTANCE,
+                type = RecordType.DISTANCE,
                 currentValue = 2,
                 targetValue = 5
             )
